store drugEffect as parameter, add option for recording population at different time intervals

+++ src/CellModel.cpp
@@ -23,7 +23,7 @@ Rcpp::List CellModel(
 	bool inheritGrowth,
 	double nG,
 	double timeIncrement,
-	double recordIncrement
+	int recordIncrement
 
 ) {
